Moved PresetSave and PresetDelete to their own file
